feat(PT-47): move Specialist type to local feature types folder
